requirements.txt
django-model-utils>=2.0
# Additional requirements go here

djangorestframework>=3.0.0
